BMI: Adapt deleteDependencyRecordsFor calls
Since the first port of BMI, the concept of extensions was introduced
into PG.  The function deleteDependencyRecordsFor received an option to
skip the dependencies of such extensions.

This patch sets the new parameter skipExtensionDeps to false.  This
resembles the original behavior.

Remark: I am not sure whether the bitmap index would be considered an
extension from today's point of view.  Currently it is not treated as
such.