BMI LOV: Update call to heap_insert
In the new prototype the negation of the former parameters 'use_wal' and
'use_fsm' is implemented as the bit flags HEAP_INSERT_SKIP_WAL and
HEAP_INSERT_SKIP_FSM in the new options parameter.  So options is now 0
to resemble the old behavior.  The last most parameter 'bistate' is new
and set to NULL to resemble the old behavior.