BMI: Replace while(true)+break with proper head
There is a search loop in _bitmap_findbitmaps which used a 'while
(true)' head and a break.  This patch changes this to a proper head
which directly uses the value which decides about leaving the loop.
This is more compact and should be easier to read.  If I did not make
any mistakes, the behavior should be unchanged.