BMI: Unpretty fix for elog code in fill_matched
Here a uint64 shall be printed.  There is no portable fixed printf code
to do that, so use some config defines to switch between %li (on 64bit
platforms) and %lli (on 32bit platforms).  It would be nicer to use the
C99 macro PRIu64 from inttypes.h but I do not know whether this is
allowed by the coding conventions.