BMI: Fix INSERT
Resurrect the old version of buf_add_tid_with_fill from Gavin Sherry's
patch (sent to the patches mailinglist on 2007-05-03) as
buf_add_tid_with_fill_immediate.  This variant does not use the HOT
buffer of the BMTIDBuffer because at the time it was written this HOT
buffer did not exist.  Nevertheless, its functionality was exactly what
was required to revive the effectiveness of insertsetbit which is the
helper of inserttuple (the non-"build" version).

insertsetbit uses a one shot BMTIDBuffer to insert a single bit into
the vector.  Therefore, the buffering is useless and the work with it
can be avoided.