Fix translation of xlogreader.c messages.
This still doesn't solve how the messages would be translated in standalone
programs. The problem is how to get the strings from xlogreader.c into
the message catalog of the stand-alone program. You could duplicate the
translation work in the standa-alone program, but probably a better approach
would be to put the xlogreader.c messages in a separate message domain
and catalog or something. But this patch doesn't try to solve that..