wal decoding: support decoding toast datums
Toast support is trivial if we don't want to reassemble if from WAL but just
read it from the respective toast table if the output plugin wants to do so. As
there are no separate visibility checks on toast tables, we can simply use the
normal HeapTupleSatisfiesToast. So don't error out if its used.

There are still arguments for reassembling toast while decoding (like being
able to up the xmin horizon earlier), but this way we support toast ok enough
for now.