BMI: Defer LOV index creation to avoid heavy WAL
When a bitmap index is created on hashable attributes, the build routine
uses a temporary hash map to map from attribute value to VMIID.
Therefore, the index on the LOV is not needed at this point in time.

This patch defers the creation of the B-tree index on the LOV until all
LOV entries have been inserted into the table.  This avoids all the
small WAL entries for the B-tree.  However, this is still a bit slower
than switching off the B-tree WAL completely, but I do not know whether
that is safe.