BMI: Fix build buffering and HOT tuple handling
This is a 80% rewrite of the HOT tuple handling code.  It aligns the HOT
buffer words to the actual bitmap words by allowing the last word of a
HOT buffer to overlap into the first word of the next HOT buffer.  Big
gaps between matching tuples are filled in block granularity.