An extension to enable testing of logical replication with isolationtester
This extension copies logical replication decoding and output code from
walsender, but alters it to return output in a more SQL-friendly fashion
(with an SRF) than the streaming format. This makes it possible to write
tests to expect specific replication output at specific times.

Some of this code is hackish, but it works.