Make init_logical_replication take a slot name and plugin options
The first piece of this patch adds a "slotname" argument to
init_logical_replication and extends that to all relevant code
(LogicalDecodingAcquireFreeSlot, test_logical_replication, etc.).
This is easy and unproblematic.

The second piece adds an option list to the init_logical_replication.
This part is untested (how can it be tested?), and doesn't work. The
grammar should probably copy some productions from the main grammar
(e.g. opt_boolean_or_string, NumericOnly), which means copying bits
of the lexer as well.

(This patch doesn't do anything with the options. They should be passed
to the pg_decode_init callback, but though the natural caller seems to
be LogicalDecodingAcquireFreeSlot, the responsibility is currently not
easily wrested from initial_snapshot_reader.)