Move plugin options from INIT_… to START_LOGICAL_REPLICATION
We're going to initialise the plugin only on START, so there's no point
passing the options to INIT. This also solves the problem of remembering
plugin options across disconnects.

To start with, option names conform to the same rules that slot and
plugin names do, and arguments must be SCONSTs. In the future, we may
have to accept boolean (and perhaps integer) arguments, but there's no
need to do it now.

    START_LOGICAL_REPLICATION 'test' 'xxxx/xxxx' ('a', 'foo' 'bar');

The options are still not handled anywhere.