Add support for in-memory out-of-line stored toast data
For some use-cases like logical decoding of wal or avoidance of repeated
detoasting of toast tuples it is useful to represent varlenas that are stored
in memory but outside of a HeapTuple. E.g. because it is impossible to allocate
a memory chunk for the whole tuple at once or because the overhead of repeated
detoasting from disk is noticeable.

For that support "indirect" toast tuples which point to another 'struct
varlena' in memory. That referenced tuple can be short, compressed or normal.

As there are no further flag bits available to discern such an indirection from
the other forms inside the original Datum we reuse the VARATT_IS_EXTERNAL
representation for this which apt because they are also
out-of-line. Unfortunately only varatt_1b_e.va_length is available to discern
toast and indirection toast tuple, so they must be different for both. A
StaticAssertStm to make sure of that is included.